# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vesta Family Logistics OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /actions:
    get:
      operationId: listActions
      tags: [actions]
      summary: List pending suggested actions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, dismissed, auto_handled]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuggestedAction"
  /actions/{id}/approve:
    post:
      operationId: approveAction
      tags: [actions]
      summary: Approve a suggested action (writes calendar event + tasks)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionApproval"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestedAction"
  /actions/{id}/dismiss:
    post:
      operationId: dismissAction
      tags: [actions]
      summary: Dismiss a suggested action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestedAction"
  /actions/{id}/edit:
    patch:
      operationId: editAction
      tags: [actions]
      summary: Edit and approve a suggested action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionEdit"
      responses:
        "200":
          description: Updated and approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestedAction"

  # ── Tasks ────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Contacts ─────────────────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List household contacts
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Rules ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, name, category]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        category:
          type: string
          enum: [escola, saude, casa, diarista, portaria, sindico, social, servicos, familia, outros]
        aliases:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    ContactInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        phone:
          type: string
        category:
          type: string
          enum: [escola, saude, casa, diarista, portaria, sindico, social, servicos, familia, outros]
        aliases:
          type: array
          items:
            type: string
        notes:
          type: string
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        category:
          type: string
        aliases:
          type: array
          items:
            type: string
        notes:
          type: string

    # ── Inbox ────────────────────────────────────────────────────
...
    SuggestedAction:
      type: object
      required: [id, inbox_item_id, category, type, title, approval_level, status, confidence]
      properties:
        id:
          type: integer
        inbox_item_id:
          type: integer
        category:
          type: string
          enum: [escola, saude, casa, social, logistica, refeicoes, servicos, outros]
        type:
          type: string
          enum: [event, task, reminder, fyi]
        title:
          type: string
        datetime:
          type: ["string", "null"]
        suggested_owner:
          type: ["string", "null"]
        approval_level:
          type: string
          enum: [soft, one_tap, explicit]
        confidence:
          type: number
        status:
          type: string
          enum: [pending, approved, dismissed, auto_handled]
        notes:
          type: ["string", "null"]
        cascade_check_needed:
          type: boolean
        workflow_tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    ActionApproval:
      type: object
      properties:
        notes:
          type: string
    ActionEdit:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        type:
          type: string
        datetime:
          type: string
        suggested_owner:
          type: string
        notes:
          type: string

    # ── Tasks ─────────────────────────────────────────────────────
...
    CalendarEvent:
      type: object
      required: [id, title, start_at, category, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: ["string", "null"]
          format: date-time
        all_day:
          type: boolean
        category:
          type: string
          enum: [escola, saude, casa, social, logistica, refeicoes, servicos, outros]
        members:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [manual, auto, pattern]
        sync_status:
          type: string
          enum: [synced, pending, failed, local]
        notes:
          type: ["string", "null"]
        workflow_tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    CalendarEventInput:
      type: object
      required: [title, start_at, category]
      properties:
        title:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        all_day:
          type: boolean
        category:
          type: string
          enum: [escola, saude, casa, social, logistica, refeicoes, servicos, outros]
        members:
          type: array
          items:
            type: string
        notes:
          type: string
    CalendarEventUpdate:
      type: object
      properties:
        title:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        all_day:
          type: boolean
        category:
          type: string
        notes:
          type: string

    # ── Rules ─────────────────────────────────────────────────────
...